refactor(shared): move pokemon data files from shared package to server
All JSON data files (pokemon.json, pokemon-encounters.json,
pokemon-evolutions.json, pokemon-gifts.json, pokemon-moves.json,
pokemon-versions.json, regional-pokedexes.json, legendary-pokemon.json,
starter-pokemon.json, trade-evolution-pokemon.json) were living in
packages/shared/data/ but were only ever consumed by server-side code
(server/features/mod.ts for app composition, scripts/sync-*.ts for
data syncing from PokeAPI).

Placing them under @make-the-pick/shared incorrectly implied they were
part of the cross-package contract shared between client and server.
More critically, it risked accidentally bundling ~5.4MB of JSON into
the client bundle if a build tool ever followed the package boundary
without careful tree-shaking.

Moving them to server/data/ makes the ownership explicit: these are
server-only assets. Updated all import and output paths in
server/features/mod.ts and all five sync scripts accordingly.
